updated readME, code and tests

+++ src/main/java/com/nadiamm/statistics/StatsController.java
@@ -3,9 +3,20 @@ package com.nadiamm.statistics;
 import java.io.IOException;
 import java.util.List;
 
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.nadiamm.statistics.model.HourlyStats;
+import com.nadiamm.statistics.model.RequestList;
+import com.nadiamm.statistics.repository.CustomerRepository;
+import com.nadiamm.statistics.repository.HourlyStatsRepository;
+import com.nadiamm.statistics.repository.RequestRepository;
+import com.nadiamm.statistics.service.StatsService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -17,7 +28,7 @@ public class StatsController {
 	@GetMapping(path="/total")
 	public ResponseEntity<Long> getRequestRepositorySize(@RequestParam String customerID, @RequestParam String tagID ) throws IOException {
 		statsService.clearRequestRepository();
-		 return ResponseEntity.ok(statsService.getTotalRequestByCustomerIDAndTagID(Integer.parseInt(customerID),Integer.parseInt(tagID)));
+		 return ResponseEntity.ok(statsService.getTotalRequestByCustomerIDAndTagID(Integer.parseInt(customerID),Long.parseLong(tagID)));
 	}	
 	
 	@GetMapping(path="/statistics")
@@ -35,5 +46,5 @@ public class StatsController {
         }
 		 return statsService.findHourlyStats();
 	}
-	
+
 }
